Changed event injection mechanism from custom to monkeyrunner
In this patch we use the ChimpChat class from chimpchat.jar to send
events to the device instead of the custom tool. This allows us to send
commands to non-rooted devices. The code is adapted to use different
keycodes compatible with monkeyrunner and to work with the libs of
platform 15.